+++ cli/src/main.rs
@@ -8,8 +8,8 @@ use track2kml_cli::{CourageousInitializer, Format};
 fn main() -> ExitCode {
     let formats: HashMap<&str, Format, RandomState> = {
         let arr = [(
-            "courageous-v0.3",
-            Format::new(CourageousInitializer, "json", "COURAGEOUS (v0.3)"),
+            "courageous-v0.4",
+            Format::new(CourageousInitializer, "json", "COURAGEOUS (v0.4)"),
         )];
 
         HashMap::from_iter(arr.into_iter())