+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 These refer to the track2kml crate versions.
+## 0.4.0
+- Update COURAGEOUS format dependency to v0.6.0+schema.0.4.0
 
 ## 0.3.0
 - Update COURAGEOUS format dependency to 0.4.0+schema.0.3.1