+++ cli/CHANGELOG.md
@@ -1,4 +1,8 @@
 ## 2.2.0
+- Update COURAGEOUS schema support to v0.4.0
+- Implemented CUAS location representation.
+
+## 2.2.0
 - Update COURAGEOUS schema support to v0.3.1
 - Display record velocity in its KML extended data balloon